+++ CurrencyService/DAL/Model/CurrencyDO.cs
@@ -6,7 +6,7 @@ namespace CurrencyService.DAL.Model
     public class CurrencyDO
     {
         public string CurrencyName { get; set; }
-        public float CurrencyValue { get; set; }
+        public double CurrencyValue { get; set; }
         public DateTime? CurrencyLastUpdate { get; set; }
         public provider CurrencyService { get; set; }
     }
